Update README.md
So this is still In the works but like more or less how it can be done. I think we should spend some time thinking very nicely about the design so that we can do it properly and efficiently.  And while thinking about the design we can work on integral aspects of the program for the time being. Like the random expression generator and how it will be designed to suit the type of difficulty whether it be selective or progressive in-game. Another thing would be the documentation that i'm working on. The other thing we need and isn't too dependent on things is the database. We need to design that and then go about seeing how a milestone system works with a game.